refactor: reorganize inquirer tool and remove deprecated files
