+++ src/lib/tools/basicTools.js
@@ -2,7 +2,7 @@ import fs from 'node:fs'
 import path from 'path'
 import { Tool } from '../Tool.js'
 import { ACTION_TAG } from '../prompt.js'
-import { inquirerPlugin } from './inquirer.js'
+import { inquirerPlugin } from './inquirer/inquirer.js'
 import cliTool from './cli.js'
 import { installSurgeonAction } from './surgeon/surgeon.js'
 import { Config } from '../config/Config.js'
