+++ src/lib/tools/inquirer/README.md
@@ -1,0 +1,11 @@
+Inquirer is a tool that helps you ask questions to the user and then perform actions based on the answers.
+
+> NOTE this is a work in progress
+
+# Requirements
+- the `silk` api
+
+# Usage
+```sh
+silk do "can you create a website for me?"
+```
